refactor: merge controller into the endpoint class
